test: move stalled-frontend backpressure test to its own file
inspect.test.ts's 'websocket > bun --inspect*' cases fail in environments
where localhost prefers ::1 (the ws client cannot reach the IPv4-bound
inspector), which is unrelated to this change but made the file unusable
as the gate's before/after target.
